Prevents memo display from overflowing.
Roughly approximate the result of equations so that they do not overflow
the display on the page. This approximation is not mathematically
accurate (i.e., it truncates decimals from the significand of the value
in E notation to fit the display), but seems sufficient for  the intent
of this project.

Also fixes a bug where negative exponents in E notation were not being
parsed properly.