Changes calculator behaviour outside assignment parameters.
The instructions call for "not [evaluating] more than a single pair of
numbers at a time". However, this behaviour felt odd and intrusive,
especially when a user pressed an operator button while the display read
"0".

Now, when the first operand is set and an operator button is pressed, it
will only evaluate if the display is not "0" as three of the four
operators do nothing or error when the second operand is `0`. If a user
really wanted to get `0` from a multiplication by zero, they can use the
`=` or `AC` buttons.